Move related types to ProjectEngineHost namespace
