+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/ExportCustomProjectEngineFactoryAttribute.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Composition;
+using Microsoft.AspNetCore.Razor.ProjectEngineHost;
 
 namespace Microsoft.CodeAnalysis.Razor;
